# apigrupo11/public/openapi.yaml
openapi: 3.0.4
info:
  title: API de Stock de Bienes y Servicios (Autenticación por ID de Usuario)
  description: |
    Servicio de gestión de stock para el subsistema de Bienes y Servicios.
    Incluye operaciones CRUD de productos, reserva y liberación de stock.
    La autorización se realiza mediante el ID de usuario explícito en las solicitudes.
  version: 1.4.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://stock.ds.frre.utn.edu.ar/v1
...
paths:
  /productos:
    get:
      tags: [Productos]
      summary: Listar productos
      parameters:
        - $ref: '#/components/parameters/PaginaActual'
        - $ref: '#/components/parameters/PorPagina'
        - $ref: '#/components/parameters/Busqueda'
        - name: categoriaId
          in: query
          schema:
            type: integer
          description: "Filtrar productos por el ID de una categoría específica."
      responses:
        '200':
          description: Lista de productos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Producto'
    post:
      tags: [Productos]
      summary: Crear un producto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductoInput'
      responses:
        '201':
          description: Producto creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoCreado'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /reservas:
    get:
      tags: [Reservas]
      summary: Listar reservas de un usuario
      parameters:
        - name: usuarioId
          in: query
          required: true
          schema:
            type: integer
          description: "ID del usuario para filtrar las reservas."
        - $ref: '#/components/parameters/PaginaActual'
        - $ref: '#/components/parameters/PorPagina'
        - name: estado
          in: query
          schema:
            type: string
            enum: [confirmado, pendiente, cancelado]
          description: "Filtrar reservas por estado"
      responses:
        '200':
          description: Lista de reservas del usuario
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReservaCompleta'
...
components:
  parameters:
    PaginaActual:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
      description: Número de página
    PorPagina:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
      description: Cantidad de resultados por página
    Busqueda:
      name: q
      in: query
      schema:
        type: string
      description: Texto de búsqueda
  schemas:
    Producto:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
          minLength: 1
          maxLength: 200
        descripcion:
          type: string
        precio:
          type: number
          format: float
          minimum: 0
        stockDisponible:
          type: integer
          minimum: 0
        pesoKg:
          type: number
          format: float
          minimum: 0
        imagenes:
          type: array
          items:
            $ref: '#/components/schemas/ImagenProducto'
          description: "Lista de imágenes del producto. Solo una puede ser marcada como principal."
        categorias:
          type: array
          items:
            $ref: '#/components/schemas/Categoria'
          nullable: true
    ProductoInput:
      type: object
      required: [nombre, precio, stockInicial]
      properties:
        nombre:
          type: string
        descripcion:
          type: string
        precio:
          type: number
          format: float
        stockInicial:
          type: integer
        pesoKg:
          type: number
          format: float
        imagenes:
          type: array
          items:
            $ref: '#/components/schemas/ImagenProducto'
        categoriaIds:
          type: array
          items:
            type: integer
          description: "Lista de IDs de las categorías a las que pertenece el producto."
      example:
        nombre: "Laptop Pro X"
        descripcion: "Una laptop potente para profesionales."
        precio: 1499.99
        stockInicial: 50
        pesoKg: 1.8
        categoriaIds: [1, 5]
...
    ProductoCreado:
      type: object
      properties:
        id:
          type: integer
        mensaje:
          type: string
      example:
        id: 101
        mensaje: "Producto creado exitosamente."
...
    ReservaCompleta:
      type: object
      properties:
        idReserva:
          type: integer
        idCompra:
          type: string
        usuarioId:
          type: integer
        estado:
          type: string
          enum: [confirmado, pendiente, cancelado]
        expiresAt:
          type: string
          format: date-time
        fechaCreacion:
          type: string
          format: date-time
        fechaActualizacion:
          type: string
          format: date-time
        productos:
          type: array
          items:
            type: object
            properties:
              idProducto:
                type: integer
              nombre:
                type: string
              cantidad:
                type: integer
              precioUnitario:
                type: number
          description: "Detalle de los productos reservados."
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: string
          nullable: true
      example:
        code: "INVALID_DATA"
        message: "Los datos proporcionados son inválidos."
        details: "El campo 'usuarioId' es requerido."
    Categoria:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
          minLength: 1
          maxLength: 100
        descripcion:
          type: string
          nullable: true
      example:
        id: 1
        nombre: "Electrónica"
        descripcion: "Dispositivos electrónicos y accesorios."
...
    ImagenProducto:
      type: object
      required: [url, esPrincipal]
      properties:
        url:
          type: string
          format: uri
          description: "URL de la imagen."
        esPrincipal:
          type: boolean
          description: "Indica si esta es la imagen principal del producto. Solo una imagen puede ser la principal."
      example:
        url: "https://example.com/images/laptop_frontal.jpg"
        esPrincipal: true
...